+++ packages/ai/core/tool/mcp/mcp-client.test.ts
@@ -1,13 +1,15 @@
+import { z } from 'zod';
+import { MCPClientError } from '../../../errors';
 import { createMCPClient } from './mcp-client';
 import { MockMCPTransport } from './mock-mcp-transport';
-import { z } from 'zod';
 import { CallToolResult } from './types';
-import { MCPClientError } from '../../../errors';
 
 const createMockTransport = vi.fn(config => new MockMCPTransport(config));
 
-vi.mock('./mcp-transport.ts', () => {
+vi.mock('./mcp-transport.ts', async importOriginal => {
+  const actual = await importOriginal<typeof import('./mcp-transport')>();
   return {
+    ...actual,
     createMcpTransport: vi.fn(config => {
       return createMockTransport(config);
     }),
@@ -28,7 +30,7 @@ describe('MCPClient', () => {
 
   it('should return AI SDK compatible tool set', async () => {
     client = await createMCPClient({
-      transport: { type: 'stdio', command: 'node', args: ['test.js'] },
+      transport: { type: 'sse', url: 'https://example.com/sse' },
     });
     const tools = await client.tools();
     expect(tools).toHaveProperty('mock-tool');
@@ -66,7 +68,7 @@ describe('MCPClient', () => {
 
   it('should return typed AI SDK compatible tool set', async () => {
     client = await createMCPClient({
-      transport: { type: 'stdio', command: 'node', args: ['test.js'] },
+      transport: { type: 'sse', url: 'https://example.com/sse' },
     });
     const tools = await client.tools({
       schemas: {
@@ -96,7 +98,7 @@ describe('MCPClient', () => {
 
   it('should not return user-defined tool if it is nonexistent', async () => {
     client = await createMCPClient({
-      transport: { type: 'stdio', command: 'node', args: ['test.js'] },
+      transport: { type: 'sse', url: 'https://example.com/sse' },
     });
     const tools = await client.tools({
       schemas: {
@@ -117,7 +119,7 @@ describe('MCPClient', () => {
         }),
     );
     client = await createMCPClient({
-      transport: { type: 'stdio', command: 'node', args: ['test.js'] },
+      transport: { type: 'sse', url: 'https://example.com/sse' },
     });
     const tools = await client.tools({
       schemas: {
@@ -141,7 +143,7 @@ describe('MCPClient', () => {
     );
 
     client = await createMCPClient({
-      transport: { type: 'stdio', command: 'node', args: ['test.js'] },
+      transport: { type: 'sse', url: 'https://example.com/sse' },
     });
 
     await expect(client.tools()).rejects.toThrow(MCPClientError);
@@ -157,7 +159,7 @@ describe('MCPClient', () => {
 
     await expect(
       createMCPClient({
-        transport: { type: 'stdio', command: 'node', args: ['test.js'] },
+        transport: { type: 'sse', url: 'https://example.com/sse' },
       }),
     ).rejects.toThrowError(MCPClientError);
   });
@@ -179,7 +181,7 @@ describe('MCPClient', () => {
 
     await expect(
       createMCPClient({
-        transport: { type: 'stdio', command: 'node', args: ['test.js'] },
+        transport: { type: 'sse', url: 'https://example.com/sse' },
       }),
     ).rejects.toThrowError(MCPClientError);
   });
@@ -189,7 +191,7 @@ describe('MCPClient', () => {
     const closeSpy = vi.spyOn(mockTransport, 'close');
     createMockTransport.mockImplementation(() => mockTransport);
     const client = await createMCPClient({
-      transport: { type: 'stdio', command: 'node', args: ['test.js'] },
+      transport: { type: 'sse', url: 'https://example.com/sse' },
     });
     await client.close();
     expect(closeSpy).toHaveBeenCalled();
@@ -197,7 +199,7 @@ describe('MCPClient', () => {
 
   it('should throw Abort Error if tool call request is aborted', async () => {
     client = await createMCPClient({
-      transport: { type: 'stdio', command: 'node', args: ['test.js'] },
+      transport: { type: 'sse', url: 'https://example.com/sse' },
     });
     const tools = await client.tools();
     const tool = tools['mock-tool'];
@@ -224,9 +226,54 @@ describe('MCPClient', () => {
     });
     createMockTransport.mockImplementation(() => mockTransport);
     client = await createMCPClient({
-      transport: { type: 'stdio', command: 'node', args: ['test.js'] },
+      transport: { type: 'sse', url: 'https://example.com/sse' },
       onUncaughtError,
     });
     expect(onUncaughtError).toHaveBeenCalled();
+  });
+
+  it('should support custom transports', async () => {
+    const mockTransport = new MockMCPTransport();
+    client = await createMCPClient({
+      transport: mockTransport,
+    });
+    const tools = await client.tools({
+      schemas: {
+        'mock-tool': {
+          parameters: z.object({
+            foo: z.string(),
+          }),
+        },
+      },
+    });
+    expect(tools).toHaveProperty('mock-tool');
+    const tool = tools['mock-tool'];
+
+    type ToolParams = Parameters<typeof tool.execute>[0];
+    expectTypeOf<ToolParams>().toEqualTypeOf<{ foo: string }>();
+
+    const result = await tool.execute(
+      { foo: 'bar' },
+      {
+        messages: [],
+        toolCallId: '1',
+      },
+    );
+
+    expectTypeOf<typeof result>().toEqualTypeOf<CallToolResult>();
+  });
+
+  it('should throw if transport is missing required methods', async () => {
+    // Because isCustomMcpTransport will return false, the client will fallback to createMcpTransport, but it will throw because the transport is invalid:
+    const invalidTransport = {
+      start: vi.fn(),
+      close: vi.fn(),
+    };
+    // @ts-expect-error - invalid transport
+    createMockTransport.mockImplementation(() => invalidTransport);
+    await expect(
+      // @ts-expect-error - invalid transport
+      createMCPClient({ transport: invalidTransport }),
+    ).rejects.toThrow();
   });
 });